+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 =================
 
 ## 0.18.54
+### is.codion.common.db
+- ConnectionPoolStatistics cleanup and refactoring.
 
 ## 0.18.53
 ### is.codion.common.db